// ui/src/generated/api/model/dtoContestPermissions.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Vertex OJ API
 * Public application endpoints and the authenticated Judge worker protocol.
 * OpenAPI spec version: 1.0
 */

export interface DtoContestPermissions {
  delete: boolean;
  edit: boolean;
  eligible: boolean;
  manageAccess: boolean;
  previewProblems: boolean;
  register: boolean;
  rejudge: boolean;
  reply: boolean;
  submit: boolean;
  transfer: boolean;
  view: boolean;
  viewJury: boolean;
}